+++ src/routes/Shared.jsx
@@ -1,0 +1,17 @@
+import Nav from '../components/SharedComponents/Nav.jsx';
+import Header from '../components/SharedComponents/Header.jsx';
+import Content from '../components/SharedComponents/Content.jsx';
+import Footer from '../components/SharedComponents/Footer.jsx';
+
+function Shared() {
+  return (
+    <>
+      <Nav />
+      <Header />
+      <Content />
+      <Footer />
+    </>
+  );
+}
+
+export default Shared;
